Modify scanlines per field, based on artifact bit
Bit 2 of $0400 will remove color artifacts (on the original machine), but it also changes the number of lines per field.
With the bit set (normal when displaying text), there are 263 lines per field; with the bit reset, there are only 262.
This fix corrects this situaiton, and makes the core act like the real machine.